[FIXES] Added more information to README and fixed a few errors.
- Added more information about the installation process;
- Try to fixed missing data packages;
- Changed string output for failed and unstranslatable warnings;
- Added "exchange" paramater support for Infobox Item; and
- Small adjustment in the example script.

+++ examples/infobox_item.py
@@ -9,10 +9,11 @@ from GlouphrieTranslator import get_infobox_item, get_rsw
 
 # Get the page and check if it exists
 page, exists = get_rsw("Zaros godsword")
-# Get the infobox item
-pagina = get_infobox_item(page)
+if exists:
+    # Get the infobox item
+    pagina = get_infobox_item(page)
 
-# Write the infobox item to a file
-with open("zgs_ptbr.txt", "w", encoding="utf-8") as f:
-    f.write(pagina)
-    f.close()
+    # Write the infobox item to a file
+    with open("zgs_ptbr.txt", "w", encoding="utf-8") as f:
+        f.write(pagina)
+        f.close()
